Azure Functions/requirements.txt
azure-functions
datetime
markdownify
azure-storage-blob
azure-identity
beautifulsoup4
urllib3
lxml